- fixed the usedcolor array's base type.
The usedcolors array which counts the number of pixels in a given color in a font used bytes as storage, so any color that just happened to have a count that is a multiple of 256 the color was considered not present.

# Conflicts:
#	src/gamedata/fonts/font.cpp
#	src/gamedata/fonts/fontinternals.h
#	src/gamedata/fonts/specialfont.cpp
#	src/gamedata/fonts/v_font.cpp
#	src/v_font.h

+++ src/v_font.h
@@ -113,7 +113,7 @@ protected:
 		const void *ranges, int total_colors, const PalEntry *palette);
 	void FixXMoves();
 
-	static int SimpleTranslation (uint8_t *colorsused, uint8_t *translation,
+	static int SimpleTranslation (uint32_t *colorsused, uint8_t *translation,
 		uint8_t *identity, double **luminosity);
 
 	int FirstChar, LastChar;
